+++ ole/mpz-ole-core/src/lib.rs
@@ -1,8 +1,13 @@
+#![deny(missing_docs, unreachable_pub, unused_must_use)]
+#![deny(unsafe_code)]
+#![deny(clippy::all)]
+
+pub mod ole;
 pub mod role;
 
-use thiserror::Error;
-
-#[derive(Debug, Error)]
+#[allow(missing_docs)]
+#[derive(Debug, thiserror::Error)]
+/// An error
 pub enum OLECoreError {
     #[error("{0}")]
     LengthMismatch(String),
