+++ ole/mpz-ole-core/src/ole/mod.rs
@@ -1,0 +1,3 @@
+//! Provides implementations of OLE with errors (OLEe) protocols.
+
+pub mod role;
