+++ ole/mpz-ole-core/src/ole/role/evaluator.rs
@@ -1,0 +1,1 @@
+pub struct OLEeEvaluator;
