+++ ole/mpz-ole-core/src/ole/role/mod.rs
@@ -1,0 +1,7 @@
+//! Provides implementations of OLEe protocols based on ROLEe.
+
+mod evaluator;
+mod provider;
+
+pub use evaluator::OLEeEvaluator;
+pub use provider::OLEeProvider;
