+++ ole/mpz-ole-core/src/ole/role/provider.rs
@@ -1,0 +1,1 @@
+pub struct OLEeProvider;
