+++ ole/mpz-ole/src/role/mod.rs
@@ -1,3 +1,3 @@
 //! Provides implementations of random OLE with errors (ROLEe) protocols.
 
-pub mod ot;
+pub mod rot;
